tests/wpt/web-platform-tests/css/work-in-progress/ttwf_sf/cshoop/medialist-creating-object-001.html

<!DOCTYPE html>
<html>
<head>
  <title>CSS Test: CSSOM MediaList Object Creation</title>
  <link rel="author" title="Ben Sheldon" href="mailto:ben@codeforamerica.org">
  <link rel="author" title="Chapman Shoop" href="mailto:chapman.shoop@gmail.com">
  <link rel="help" href="https://drafts.csswg.org/cssom/#creating-a-medialist-object">
  <meta name="flags" content="dom">
  <meta name="assert" content="Create a new MediaList object.">
  <script src="/resources/testharness.js" type="text/javascript"></script>
  <script src="/resources/testharnessreport.js" type="text/javascript"></script>
</head>

<body>

  <noscript>Test not run - javascript required.</noscript>
  <div id="log"></div>

  <script type="text/javascript">
    test(function() {
      var my_media_list = new MediaList();
      my_media_list.mediaText = 'all';
      assert_true(my_media_list instanceof MediaList);
    }, "medialist_spec_object_creation",
    { assert: ["Can create MediaList object as described in the spec."] });
  </script>
</body>
</html>
